requirements-visualizacao.txt
numpy==1.26.4
PyOpenGL==3.1.7
glfw==2.7.0

